FitrackFrontend/src/app/services/admin.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpHeaders } from '@angular/common/http';

import { AdminService } from './admin.service';

describe('AdminService', () => {
  let service: AdminService;
  let httpMock: HttpTestingController;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [AdminService]
    });
    service = TestBed.inject(AdminService);
    httpMock = TestBed.inject(HttpTestingController);
  });

  afterEach(() => {
    httpMock.verify();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  it('should make a POST request with the correct headers and body', () => {
    const adminData = { fullName: 'Admin User', email: 'admin@example.com', password: 'password123' };
    const token = 'dummy-token';
    localStorage.setItem('token', token);

    service.createAdmin(adminData).subscribe();

    const req = httpMock.expectOne('http://localhost:8080/admins');
    expect(req.request.method).toBe('POST');
    expect(req.request.headers.get('Authorization')).toBe(`Bearer ${token}`);
    expect(req.request.body).toEqual(adminData);

    req.flush({ message: 'Admin created successfully' });
  });
});
